pyproject.toml
[project]
name = "taniclaw"
version = "1.0.0"
description = "Lightweight autonomous agriculture skill — Food Security Agent for Everyone"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "apscheduler>=3.10.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "groq>=0.9.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
test = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0", "httpx>=0.27.0"]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0", "httpx>=0.27.0"]

[project.scripts]
taniclaw = "taniclaw.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["taniclaw*"]
namespaces = false

[tool.setuptools.package-data]
taniclaw = [
    "rules/*.yaml",
    "knowledge/**/*.yaml",
    "knowledge/**/*.yml",
    "frontend/**/*"
]

[tool.uv]
managed = true
package = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
